Set image preview CSP
Using data url's is probably the least worst option.

Nonces could be used be only up to a fixed amount and would require architecture changes to get them into the appropriate js

We could show a loading spinner when uploading and fetch the image from the server but that is inefficient for bandwidth and latency reasons. Plus there are issues with non-committed images.

Custom drawing the image from JS onto a canvas is also possible, but complex, non-semantic and likely non-performant